+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
++ Update `--paging` option to `--no-pager` and honor $PAGER to be more Unix-like
+
 ## 0.1.2dev1 (Mar 7, 2024)
 
 + Fixed word wrap in code cells. Closes [Text wrapping for code cells #1](https://github.com/tnwei/nbread/issues/1)